lib/tst_test: Cast rlim_t to long long unsigned int
That fixes warnings on musl:

tst_test.c:1156:76: warning: format '%lu' expects argument of type 'long
unsigned int', but argument 6 has type 'rlim_t' {aka 'long long unsigned
int'} [-Wformat=]
 1156 |         tst_res_(file, lineno, TINFO, "Set ulimit resource: %d rlim_cur: %lu rlim_max: %lu",
      |                                                                          ~~^
      |                                                                            |
      |                                                                            long unsigned int

+++ lib/tst_test.c
@@ -1153,8 +1153,9 @@ static void set_ulimit_(const char *file, const int lineno, const struct tst_uli
 	if (conf->rlim_cur > rlim.rlim_max)
 		rlim.rlim_max = conf->rlim_cur;
 
-	tst_res_(file, lineno, TINFO, "Set ulimit resource: %d rlim_cur: %lu rlim_max: %lu",
-		conf->resource, rlim.rlim_cur, rlim.rlim_max);
+	tst_res_(file, lineno, TINFO, "Set ulimit resource: %d rlim_cur: %llu rlim_max: %llu",
+		conf->resource, (long long unsigned int)rlim.rlim_cur,
+		(long long unsigned int)rlim.rlim_max);
 
 	safe_setrlimit(file, lineno, conf->resource, &rlim);
 }
